backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
ruff
pre-commit
pytest
pytest-cov
pytest-dotenv
poethepoet
redis
huggingface-hub